refactor: separate command execution
